everything done, next path and url in vm

+++ srcs/requirements/wordpress/config/start_wordpress.sh
@@ -1,4 +1,17 @@
 #!/bin/sh
+
+check_admin_username() {
+    forbidden_substrings="admin Admin administrator Administrator ADMIN"
+    for substring in $forbidden_substrings; do
+        if echo "$WORDPRESS_ADMIN_USER" | grep -i -q "$substring"; then
+            echo "Error: The administrator's username cannot contain 'admin/Admin' or 'administrator/Administrator' in any form."
+            exit 1
+        fi
+    done
+}
+
+check_admin_username
+
 if [ ! -f /var/www/html/wp-config.php ]; then
     echo "WordPress not found. Downloading and installing..."
     rm -rf /var/www/html/*
@@ -37,7 +50,7 @@ if [ ! -f /var/www/html/wp-config.php ]; then
 
     echo "Installing WordPress..."
     wp core install \
-        --url="$DOMAIN_NAME" \
+        --url="http://localhost" \
         --title="Inception" \
         --admin_user="$WORDPRESS_ADMIN_USER" \
         --admin_password="$WORDPRESS_ADMIN_PASSWORD" \
@@ -65,6 +78,9 @@ else
     echo "WordPress already installed."
 fi
 
+echo "Updating siteurl and home options..."
+wp option update siteurl http://localhost --allow-root --path=/var/www/html/
+wp option update home http://localhost --allow-root --path=/var/www/html/
 # echo "Configuring PHP-FPM to listen on 0.0.0.0:9000..."
 sed -i 's/listen = .*/listen = 0.0.0.0:9000/' /etc/php/7.4/fpm/pool.d/www.conf
